Patch to switch between mailR library and blastula.
mailR depends on Java libraries being previously installed on the
system, and while that is possible it creates an unnecessary point
of failure.  mailR used TLS security, which is the current state
of the art but isn't necessary for a feedback control on a scientific
website.  Blastula uses SSL security which is a step down but still
perfectly valid for our uses.